EPMRPP-115034 || Refactor command execution in ExecuteIntegrationHandlerImpl and related classes
