+++ src/main/java/com/epam/reportportal/extension/command/ExtensionCommand.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright (C) 2026 EPAM Systems
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.epam.reportportal.extension.command;
 
 import com.epam.reportportal.api.model.PluginCommandRQ;
@@ -13,7 +29,7 @@ public interface ExtensionCommand<T> extends NamedPluginCommand {
    * @return Result
    */
   default T executeCommand(PluginCommandRQ pluginCommandRq) {
-    return null;
+    throw new UnsupportedOperationException("Command does not support execution without an integration");
   }
 
 
@@ -25,7 +41,7 @@ public interface ExtensionCommand<T> extends NamedPluginCommand {
    * @return Result
    */
   default T executeCommand(Integration integration, PluginCommandRQ pluginCommandRq) {
-    return null;
+    throw new UnsupportedOperationException("Command does not support execution with an integration");
   }
 
 }
