--- src/main/java/com/epam/reportportal/extension/command/OrganizationManagerExtensionCommand.java
@@ -1,40 +1,0 @@
-/*
- * Copyright 2025 EPAM Systems
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.epam.reportportal.extension.command;
-
-import com.epam.reportportal.base.infrastructure.persistence.dao.organization.OrganizationRepositoryCustom;
-import com.epam.reportportal.base.infrastructure.persistence.dao.organization.OrganizationUserRepository;
-import com.epam.reportportal.base.infrastructure.persistence.entity.organization.OrganizationRole;
-import com.epam.reportportal.base.infrastructure.persistence.entity.user.OrganizationUser;
-import com.epam.reportportal.base.infrastructure.rules.exception.ErrorType;
-import com.epam.reportportal.base.infrastructure.rules.exception.ReportPortalException;
-
-public abstract class OrganizationManagerExtensionCommand<T> extends OrganizationMemberExtensionCommand<T> {
-
-  protected OrganizationManagerExtensionCommand(OrganizationRepositoryCustom organizationRepository,
-      OrganizationUserRepository organizationUserRepository) {
-    super(organizationRepository, organizationUserRepository);
-  }
-
-  @Override
-  protected void validateOrganizationRole(OrganizationUser organizationUser) {
-    if (!organizationUser.getOrganizationRole().sameOrHigherThan(OrganizationRole.MANAGER)) {
-      throw new ReportPortalException(ErrorType.ACCESS_DENIED, "Organization Manager role required.");
-    }
-  }
-
-}
